cleanup(rewrite) buffer and classic reorganization
